@@ -15,121 +15,69 @@
   ],
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/Firestitch/fs-form/issues"
+    "url": "https://github.com/Firestitch/ngx-form/issues"
   },
   "main": "package/index.js",
   "typings": "package/index.d.ts",
   "scripts": {
-    "package": "cd tools && npm run package",
-    "package:publish": "cd tools && npm run package:publish",
-    "demo:build": "cd tools && npm run demo:build",
-    "serve": "cd tools && npm run serve"
+    "package": "cd tools && npm run cli:package",
+    "package:test": "cd tools && npm run cli:package:test",
+    "package:publish": "cd tools && npm run cli:package:publish",
+    "demo:build": "cd tools && npm run cli:demo",
+    "serve": "cd tools && npm run cli:serve"
   },
   "devDependencies": {
-    "@angular/cdk": "^5.0.0",
-    "@angular/common": "^5.1.1",
-    "@angular/compiler": "^5.1.1",
-    "@angular/compiler-cli": "^5.1.1",
-    "@angular/core": "^5.1.1",
-    "@angular/forms": "^5.1.1",
-    "@angular/http": "^5.1.1",
-    "@angular/material": "^5.0.1",
-    "@angular/flex-layout": "2.0.0-beta.10-4905443",
-    "@angular/animations": ">=5.0.0",
-    "@angular/platform-browser": "^5.1.1",
-    "@angular/platform-browser-dynamic": "^5.1.1",
-    "@compodoc/compodoc": "^1.0.5",
-    "@types/googlemaps": "^3.30.2",
-    "@types/jasmine": "2.5.53",
-    "@types/node": "^6.0.94",
-    "@firestitch/datepicker": "^0.3.4",
-    "@firestitch/checkboxgroup": "^0.1.20",
-    "@firestitch/radiogroup": "^0.1.9",
-    "@firestitch/autocomplete-chips": "^0.1.2",
-    "@firestitch/example": "^0.3.0",
-    "@firestitch/iframe": "^0.1.5",
-    "@firestitch/account-picker": "^0.2.2",
-    "iscroll": "^5.2.0",
-    "moment-timezone": "^0.5.14",
-    "moment-range": "^3.1.1",
-    "hammerjs": "^2.0.8",
-    "angular-in-memory-web-api": "^0.3.2",
-    "codelyzer": "~3.2.0",
-    "concurrently": "^3.4.0",
-    "core-js": "^2.5.3",
-    "del": "^2.2.2",
-    "gulp": "^3.9.1",
-    "gulp-rename": "^1.2.2",
-    "gulp-rollup": "^2.15.0",
-    "gulp-replace": "^0.6.1",
-    "inquirer": "^4.0.1",
-    "highlight.js": "^9.9.0",
-    "ucfirst": "^1.0.0",
-    "gulp-regex-rename": "^0.1.0",
-    "gulp-babel": "^7.0.0",
-    "babel-core": "^6.26.0",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "^1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "lite-server": "^2.3.0",
-    "node-sass": "^4.5.2",
-    "node-sass-tilde-importer": "^1.0.0",
-    "node-watch": "^0.5.2",
-    "protractor": "~5.1.2",
-    "rollup": "^0.49.3",
-    "run-sequence": "^1.2.2",
-    "rxjs": "^5.5.5",
-    "systemjs": "^0.20.12",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.4.2",
-    "zone.js": "^0.8.14",
-    "webpack": "^3.10.0",
-    "webpack-dev-server": "^2.11.0",
-    "webpack-dll-bundles-plugin": "^1.0.0-beta.5",
-    "webpack-merge": "^4.1.1",
-    "webpack-node-externals": "^1.6.0",
-    "webpack-shell-plugin": "^0.5.0",
-    "html-webpack-plugin": "^2.30.1",
-    "css-loader": "^0.28.8",
-    "css-to-string-loader": "^0.1.3",
-    "directory-tree-webpack-plugin": "^0.2.0",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "copy-webpack-plugin": "^4.3.1",
-    "raw-loader": "^0.5.1",
-    "angular2-template-loader": "^0.6.2",
-    "awesome-typescript-loader": "^3.4.1",
-    "angular-router-loader": "^0.8.2",
-    "postcss-cssnext": "^3.0.2",
-    "postcss-flexbugs-fixes": "^3.2.0",
-    "postcss-loader": "^2.0.10",
-    "resolve-url-loader": "^2.2.1",
-    "sass-loader": "^6.0.6",
-    "to-string-loader": "^1.1.5",
-    "style-loader": "^0.19.1",
-    "npm-run-all": "^4.1.2",
+    "@angular-devkit/build-angular": "^0.12.2",
+    "@angular-devkit/build-ng-packagr": "^0.12.2",
+    "@angular/animations": "^7.2.1",
+    "@angular/cdk": "^7.2.1",
+    "@angular/cli": "^7.2.2",
+    "@angular/common": "^7.2.1",
+    "@angular/compiler": "^7.2.1",
+    "@angular/compiler-cli": "^7.2.1",
+    "@angular/core": "^7.2.1",
+    "@angular/flex-layout": "^7.0.0-beta.23",
+    "@angular/forms": "^7.2.1",
+    "@angular/http": "^7.2.1",
+    "@angular/material": "^7.2.1",
+    "@angular/platform-browser": "^7.2.1",
+    "@angular/platform-browser-dynamic": "^7.2.1",
+    "@compodoc/compodoc": "^1.1.7",
+    "@firestitch/account-picker": "^0.3.0",
+    "@firestitch/autocomplete-chips": "^0.2.0",
+    "@firestitch/checkboxgroup": "^0.2.0",
+    "@firestitch/common": "^0.5.1",
+    "@firestitch/datepicker": "^0.5.0",
+    "@firestitch/example": "^0.5.2",
+    "@firestitch/iframe": "^0.2.0",
+    "@firestitch/message": "^0.2.0",
+    "@firestitch/radiogroup": "^0.2.0",
+    "codelyzer": "^4.5.0",
     "cpx": "^1.5.0",
-    "replace": "^0.3.0",
-    "replace-in-file": "^3.4.0",
-    "lodash": "^4.17.5",
-    "@firestitch/common": "^0.2.24",
-    "@firestitch/message": "^0.1.18",
-    "ngx-toastr": "^8.7.3",
-    "filemanager-webpack-plugin": "^1.0.26"
+    "hammerjs": "^2.0.8",
+    "iscroll": "^5.2.0",
+    "moment-range": "^4.0.1",
+    "moment-timezone": "^0.5.23",
+    "ng-packagr": "^4.6.0",
+    "npm-run-all": "^4.1.5",
+    "rxjs": "^6.3.3",
+    "tsickle": "^0.34.0",
+    "tslint": "^5.12.1",
+    "typescript": "^3.1.6",
+    "webpack": "^4.29.0",
+    "webpack-dev-server": "^3.1.14",
+    "zone.js": "^0.8.28"
   },
   "engines": {
     "node": ">=6.0.0"
   },
   "peerDependencies": {
-    "rxjs": ">=5.1.0",
+    "rxjs": ">=6.3.0",
     "zone.js": ">=0.8.4",
-    "lodash": ">=4.0.0",
-    "@firestitch/common": ">=0.2.24",
+    "@firestitch/common": ">=0.5.0",
     "moment-timezone": ">=0.5.10"
+  },
+  "dependencies": {
+    "lodash-es": "^4.17.0"
   }
 }
