@@ -1,7 +1,12 @@
 document.addEventListener('DOMContentLoaded', function() {
     var $tabSource = document.querySelector('#source-tab'),
-        $prismPre = document.querySelector('pre');
-    if ($prismPre) {
+        $tabInfo = document.querySelector('#info-tab'),
+        $tabReadme = document.querySelector('#readme-tab'),
+        $tabTemplate = document.querySelector('#templateData-tab'),
+        $tabTree = document.querySelector('#tree-tab'),
+        $tabExample = document.querySelector('#example-tab'),
+        $prismPre = document.querySelector('pre.compodoc-sourcecode');
+    if ($tabSource && $prismPre) {
         $prismCode = $prismPre.querySelector('code'),
         $content = document.querySelector('.content'),
         prismLinks = document.querySelectorAll('.link-to-prism')
@@ -25,5 +30,29 @@ document.addEventListener('DOMContentLoaded', function() {
                 $content.scrollTop = top;
             }, 500);
         };
+
+        window.onhashchange = function(event) {
+            switch (window.location.hash) {
+                case '':
+                case '#info':
+                    $tabInfo.click();
+                    break;
+                case '#readme':
+                    $tabReadme.click();
+                    break;
+                case '#source':
+                    $tabSource.click();
+                    break;
+                case '#template':
+                    $tabTemplate.click();
+                    break;
+                case '#dom-tree':
+                    $tabTree.click();
+                    break;
+                case '#example':
+                    $tabExample.click();
+                    break;
+            }
+        }
     }
 });
